requirements.txt
matplotlib==3.5.3
numpy==1.23.2
pandas==1.4.4
Pillow==9.2.0
pytz==2022.2.1
scikit_learn==1.1.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
